rename to toNodeHandler
