+++ test/app.test.ts
@@ -1,7 +1,7 @@
 import { Readable, Transform } from 'stream'
 import supertest, { SuperTest, Test } from 'supertest'
 import { describe, it, expect, beforeEach } from 'vitest'
-import { createApp, nodeHandler, App, eventHandler, nodeEventHandler } from '../src'
+import { createApp, toNodeHandler, App, eventHandler, nodeEventHandler } from '../src'
 
 describe('app', () => {
   let app: App
@@ -9,7 +9,7 @@ describe('app', () => {
 
   beforeEach(() => {
     app = createApp({ debug: false })
-    request = supertest(nodeHandler(app))
+    request = supertest(toNodeHandler(app))
   })
 
   it('can return JSON directly', async () => {
